Extract ValidationMessage class
