+++ articles/ds/components/grid/index.asciidoc
@@ -587,6 +587,11 @@ include::{root}/frontend/demo/component/grid/grid-buffered-inline-editor.ts[prei
 ----
 include::{root}/src/main/java/com/vaadin/demo/component/grid/GridBufferedInlineEditor.java[render,tags=snippet,indent=0,group=Java]
 ----
+
+[source,java]
+----
+include::{root}/src/main/java/com/vaadin/demo/component/grid/ValidationMessage.java[group=Java]
+----
 --
 
 === Non-Buffered
@@ -604,6 +609,11 @@ include::{root}/frontend/demo/component/grid/grid-unbuffered-inline-editor.ts[pr
 [source,java]
 ----
 include::{root}/src/main/java/com/vaadin/demo/component/grid/GridUnbufferedInlineEditor.java[render,tags=snippet,indent=0,group=Java]
+----
+
+[source,java]
+----
+include::{root}/src/main/java/com/vaadin/demo/component/grid/ValidationMessage.java[group=Java]
 ----
 --
