+++ src/main/java/com/vaadin/demo/component/grid/GridBufferedInlineEditor.java
@@ -3,12 +3,9 @@ package com.vaadin.demo.component.grid;
 import com.vaadin.demo.DemoExporter; // hidden-source-line
 import com.vaadin.demo.domain.DataService;
 import com.vaadin.demo.domain.Person;
-import com.vaadin.flow.component.HasText;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.grid.Grid;
 import com.vaadin.flow.component.grid.editor.Editor;
-import com.vaadin.flow.component.html.Span;
-import com.vaadin.flow.component.icon.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
@@ -100,35 +97,6 @@ public class GridBufferedInlineEditor extends VerticalLayout {
         getThemeList().add("spacing-s");
         add(grid, firstNameValidationMessage, lastNameValidationMessage,
                 emailValidationMessage);
-    }
-
-    private static class ValidationMessage extends HorizontalLayout
-            implements HasText {
-
-        private final Span span = new Span();
-
-        public ValidationMessage() {
-            setVisible(false);
-            setAlignItems(Alignment.CENTER);
-            getStyle().set("color", "var(--lumo-error-text-color)");
-            getThemeList().clear();
-            getThemeList().add("spacing-s");
-
-            Icon icon = VaadinIcon.EXCLAMATION_CIRCLE_O.create();
-            icon.setSize("16px");
-            add(icon, span);
-        }
-
-        @Override
-        public String getText() {
-            return span.getText();
-        }
-
-        @Override
-        public void setText(String text) {
-            span.setText(text);
-            this.setVisible(text != null && !text.isEmpty());
-        }
     }
 
     public static class Exporter extends // hidden-source-line
