+++ src/main/java/com/vaadin/demo/component/grid/GridUnbufferedInlineEditor.java
@@ -5,13 +5,8 @@ import com.vaadin.demo.domain.DataService;
 import com.vaadin.demo.domain.Person;
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.Focusable;
-import com.vaadin.flow.component.HasText;
 import com.vaadin.flow.component.grid.Grid;
 import com.vaadin.flow.component.grid.editor.Editor;
-import com.vaadin.flow.component.html.Span;
-import com.vaadin.flow.component.icon.Icon;
-import com.vaadin.flow.component.icon.VaadinIcon;
-import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.textfield.EmailField;
 import com.vaadin.flow.component.textfield.TextField;
@@ -98,35 +93,6 @@ public class GridUnbufferedInlineEditor extends VerticalLayout {
             Editor<Person> editor) {
         textField.getElement().addEventListener("keydown", e -> editor.cancel())
                 .setFilter("event.code === 'Escape'");
-    }
-
-    private static class ValidationMessage extends HorizontalLayout
-            implements HasText {
-
-        private final Span span = new Span();
-
-        public ValidationMessage() {
-            setVisible(false);
-            setAlignItems(Alignment.CENTER);
-            getStyle().set("color", "var(--lumo-error-text-color)");
-            getThemeList().clear();
-            getThemeList().add("spacing-s");
-
-            Icon icon = VaadinIcon.EXCLAMATION_CIRCLE_O.create();
-            icon.setSize("16px");
-            add(icon, span);
-        }
-
-        @Override
-        public String getText() {
-            return span.getText();
-        }
-
-        @Override
-        public void setText(String text) {
-            span.setText(text);
-            this.setVisible(text != null && !text.isEmpty());
-        }
     }
 
     public static class Exporter extends // hidden-source-line
